+++ index.html
@@ -33,6 +33,7 @@
       </div>
     </div>
 
+    <div id="loader"></div>
     <script type="module" src="./src/main.ts"></script>
   </body>
 </html>
