+++ src/component/feature/MovieLayout.ts
@@ -1,10 +1,8 @@
 import MovieList from "../common/MovieList.js";
-import Button from "../common/Button.js";
 import hideskeleton from "../../util/hideskeleton.js";
 import {MovieData,MovieState} from "../../../types/movieDataType";
 import createSkeletonData from "../../util/createSkeletonData.js";
 import { getElement } from "../../util/utils.js";
-import { MOVIE_COUNT_PER_PAGE } from "../../constant/constant.js";
 import Modal from "../common/Modal.js";
 import MovieDetail from "../common/MovieDetail.js";
 import { fetchDetailMovie } from "../../api/fetch.js";
@@ -16,7 +14,6 @@ class MovieLayout {
         this.#state = {
             title: "지금 인기 있는 영화",
             eventName: "readMoreMovieList",
-            isPossibleMore: movieData.length === MOVIE_COUNT_PER_PAGE,
             movieData,
         }
         this.render();
@@ -53,7 +50,6 @@ class MovieLayout {
             <div id="movieListContainer">
                 ${MovieList(this.#state.movieData).template().outerHTML}
             </div>
-            ${this.#state.isPossibleMore? Button({content: '더보기', eventName: this.#state.eventName, type: 't', width:'100%'}) : ""}
         `}
     
     render(){
@@ -64,7 +60,6 @@ class MovieLayout {
         const movieListEl = getElement(".thumbnail-list");
         movieListEl.addEventListener("click", async(event) => {  
         const clickedItem = event.target.closest("li");
-        if(!clickedItem) return
         const {poster_path, title, vote_average, release_date, genres, overview} = await fetchDetailMovie(clickedItem.id)
           Modal(`${clickedItem.id}modal`,MovieDetail({poster_path,title, vote_average, release_date, genres, overview}))
           Modal.open(`${clickedItem.id}modal`)
@@ -73,18 +68,20 @@ class MovieLayout {
     
 
     newMovieListRender(dataList:MovieData[]) {
-        const ul = MovieList(dataList).template();
-        getElement('#movieListContainer')?.appendChild(ul);
-        const movieListEls = document.querySelectorAll(".thumbnail-list");
+        if(dataList.length!==0){
+            const ul = MovieList(dataList).template();
+            getElement('#movieListContainer')?.appendChild(ul);
+            const movieListEls = document.querySelectorAll(".thumbnail-list");
 
-        movieListEls.forEach(movieListEl => {
-           movieListEl.addEventListener("click", async(event) => {  
-               const clickedItem = event.target.closest("li");
-               const {poster_path, title, vote_average, release_date, genres, overview} = await fetchDetailMovie(clickedItem.id)
-                 Modal(`${clickedItem.id}modal`,MovieDetail({poster_path,title, vote_average, release_date, genres, overview}))
-                 Modal.open(`${clickedItem.id}modal`)
-                });
-         });
+            movieListEls.forEach(movieListEl => {
+            movieListEl.addEventListener("click", async(event) => {  
+                const clickedItem = event.target.closest("li");
+                const {poster_path, title, vote_average, release_date, genres, overview} = await fetchDetailMovie(clickedItem.id)
+                    Modal(`${clickedItem.id}modal`,MovieDetail({poster_path,title, vote_average, release_date, genres, overview}))
+                    Modal.open(`${clickedItem.id}modal`)
+                    });
+            })
+        }
     }
 
 }
